lsh/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
DateTime==5.0
deid==0.3.21
dicom-wsi==1.0.3
distro-info==1.0
Django==4.1.6
dnspython==2.3.0
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.0
fhirclient==4.1.0
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.10.2
iniconfig==2.0.0
isodate==0.6.1
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.7.0
mercurial==5.6.1
minio==7.1.13
multidict==6.0.4
numba==0.56.4
numcodecs==0.11.0
numpy==1.23.5
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pkgconfig==1.5.5
plotly==5.13.0
pluggy==1.0.0
pycparser==2.21
pycryptodome==3.17
pycurl==7.43.0.6
pydantic==1.10.5
pydicom==2.3.1
PyGObject==3.38.0
pyorthanc==1.11.4
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyTurboJPEG==1.7.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tifffile==2022.10.10
tiffile==2018.10.18
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.8.2
zipp==3.13.0
zope.interface==5.5.2
